requirements.txt
python-telegram-bot==20.4
uvicorn==0.22.0
starlette==0.37.0